Add cache.InitForTest for previous cache.NewForTest.  Now, the cache.NewForTest will not change global store and work as stateless as expected.
